load data into component to list all menters
