+++ client/src/components/Routes.js
@@ -4,7 +4,6 @@ import MentorPage from './pages/MentorPage';
 import InternshipPage from './pages/InternshipPage'
 import CoursePage from './pages/CoursePage'
 import CoursePages from './pages/CoursePages'
-import TablesPage from './pages/TablesPage'
 
 import Attendance from './pages/Attendance'
 import Login from './pages/LoginPage'
@@ -19,7 +18,6 @@ class Routes extends React.Component {
         <Route path='/internship' exact component={InternshipPage}  />
         <Route path='/course' exact component={CoursePage}  />
         <Route path='/courses' exact component={CoursePages}  />
-        <Route path='/tables' exact component={TablesPage}  />
 
         <Route path='/attendance' exact component={Attendance}  />
         <Route path='/login' exact component={Login}  />
