+++ client/src/components/pages/sections/AddMentorForm.js
@@ -30,7 +30,7 @@ class AddMentorForm extends Component {
 
     switch(fieldName) {
       case 'email':
-        emailValid = value.match(/^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$/i);
+        emailValid = value.notmatch(/^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$/i);
         fieldValidationErrors.email = emailValid ? '' : ' is invalid';
         break;
       case 'password':
